GPX-580: removed url-rewriter
used builtin query write capabilities.
